move Join into Appear::Util
